+++ src/utils/parameters/templateParameters.ts
@@ -1,14 +1,17 @@
 import { LOOP_LIMIT } from '../../config';
-import { RESERVED_PARAMETER_MISSING_VALUE } from '../../constants';
-import { RESERVED_PARAMETER_RESTRICTED } from '../../constants';
+import { RESERVED_PARAMETER_MISSING_VALUE, RESERVED_PARAMETER_RESTRICTED } from '../../constants';
 import { LimitReachedError } from '../../errors/LimitReachedError';
 import { PipelineExecutionError } from '../../errors/PipelineExecutionError';
 import { UnexpectedError } from '../../errors/UnexpectedError';
-import type { Parameters } from '../../types/typeAliases';
-import type { string_template } from '../../types/typeAliases';
+import type { Parameters, string_parameter_name, string_template } from '../../types/typeAliases';
+import { really_unknown } from '../organization/really_unknown';
+import { valueToString } from './valueToString';
 
 /**
  * Replaces parameters in template with values from parameters object
+ *
+ * Note: This function is not places strings into string,
+ *       It's more complex and can handle this operation specifically for LLM models
  *
  * @param template the template with parameters in {curly} braces
  * @param parameters the object with parameters
@@ -16,7 +19,10 @@ import type { string_template } from '../../types/typeAliases';
  * @throws {PipelineExecutionError} if parameter is not defined, not closed, or not opened
  * @public exported from `@promptbook/utils`
  */
-export function templateParameters(template: string_template, parameters: Parameters): string {
+export function templateParameters(
+    template: string_template,
+    parameters: Record<string_parameter_name, really_unknown>,
+): string {
     for (const [parameterName, parameterValue] of Object.entries(parameters)) {
         if (parameterValue === RESERVED_PARAMETER_MISSING_VALUE) {
             throw new UnexpectedError(`Parameter \`{${parameterName}}\` has missing value`);
@@ -60,7 +66,7 @@ export function templateParameters(template: string_template, parameters: Parame
             throw new PipelineExecutionError(`Parameter \`{${parameterName}}\` is not defined`);
         }
 
-        parameterValue = parameterValue.toString();
+        parameterValue = valueToString(parameterValue);
 
         if (parameterValue.includes('\n') && /^\s*\W{0,3}\s*$/.test(precol)) {
             parameterValue = parameterValue